Fix arrays by hand
Claude tried to make a script to do the changes, but it turned into a
large problem set of broken files. It turned out there was a bug in
the existing code due to arrays not being {} which was lurking for
years. Putting the braces around showed how this had compiled by
accident but would not have run correctly.

Signed-off-by: Stephen Smoogen <ssmoogen@redhat.com>